refactor: remove duplicacoes FiltersBar
