+++ src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/FiltersBar/consts.ts
@@ -7,3 +7,9 @@ export const referenceOrder = [
     "Entre 55 e 65 anos",
     "Mais de 65 anos",
 ];
+
+export const clearFiltersArgs = {
+    iconActive: "https://media.graphassets.com/1EOGJH6TvSMqTrjigY1g",
+    iconInactive: "https://media.graphassets.com/x37RkcUrTH6G50ganj9d",
+    label: "Limpar todos os filtros",
+};
