+++ src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/FiltersBar/modules/CoapsFiltersBar/__tests__/logic.test.ts
@@ -1,5 +1,5 @@
 import * as z from "zod/v4";
-import { onlyValidFilterValues, toHtmlSelectOptions } from "../logic";
+import { onlyValidFilterValues, toHtmlSelectOptions } from "../../../logic";
 
 // Mocks
 const mockPatientStatusEnum = z.enum(["Em dia", "A fazer"]);
