+++ src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/FiltersBar/modules/CoapsFiltersBar/logic.ts
@@ -1,45 +1,24 @@
-import type {
-    ConditionIdentifiedBy,
-    PatientAgeRange,
-    PatientStatus,
-    CommunityHealthWorker,
-} from "@/features/acf/shared/diabetes/model";
 import type * as schema from "@/features/acf/shared/diabetes/schema";
-import type * as z from "zod/v4";
-import { referenceOrder } from "./consts";
-import { toSelectConfigsCoeq } from "../CoeqFiltersBar/logic";
+import {
+    sortedOptions,
+    toHtmlSelectOptions,
+    toSelectConfigsShared,
+} from "../../logic";
+import type { SelectConfig } from "../common/SelectConfig";
 
 //TODO: Existem funcoes duplicadas entre CoapsFiltersBar e CoeqFiltersBar, podemos criar uma camada de abstração para elas
 
-export type FilterOptions =
-    | Array<CommunityHealthWorker>
-    | Array<PatientStatus>
-    | Array<ConditionIdentifiedBy>
-    | Array<PatientAgeRange>;
-
-export type SelectConfig = {
-    id: string;
-    label: string;
-    options: Array<HtmlSelectOption>;
-    isMultiSelect: boolean;
-    width: string;
-};
-
-export type HtmlSelectOption = {
-    value: string;
-    label: string;
-};
-
-export const sortedOptions = (
-    a: HtmlSelectOption,
-    b: HtmlSelectOption
-): number => referenceOrder.indexOf(a.label) - referenceOrder.indexOf(b.label);
+// export type FilterOptions =
+//     | Array<CommunityHealthWorker>
+//     | Array<PatientStatus>
+//     | Array<ConditionIdentifiedBy>
+//     | Array<PatientAgeRange>;
 
 export const toSelectConfigsCoaps = (
     filtersValues: schema.CoapsFilters
 ): Array<SelectConfig> => {
     return [
-        ...toSelectConfigsCoeq(filtersValues),
+        ...toSelectConfigsShared(filtersValues),
         {
             options: toHtmlSelectOptions(filtersValues.careTeamName).sort(
                 sortedOptions
@@ -50,22 +29,4 @@ export const toSelectConfigsCoaps = (
             width: "330px",
         },
     ];
-};
-
-export const onlyValidFilterValues = <TFilterValue>(
-    filterValues: Array<TFilterValue>,
-    schema: z.ZodType
-): Array<TFilterValue> => {
-    return filterValues.filter(
-        (filterValue: TFilterValue) => schema.safeParse(filterValue).success
-    );
-};
-
-export const toHtmlSelectOptions = (
-    filterValues: FilterOptions
-): Array<HtmlSelectOption> => {
-    return filterValues.map<HtmlSelectOption>((item) => ({
-        value: item,
-        label: item,
-    }));
 };
